ref(replays): Extract FiltersGrid for replay details tabs to share (#42215)
Noticed some shared code, and a poorly named variable, so i pulled it
out.
